Fixed typos, spelling and language
Hi there - tbh. I'm only fixing this because I'm looking for a way to contact you - because I cannot get pocketvox to run on my Linux Mint 18.3, it dies with "assertion failed: (vader)" (because vader is no longer a part of that gstreamer plugin). Do you have a working version at home?